Moved DevTest cases to separate component. Reformatting of appState
