+++ example/dapp-react/src/components/DevTests/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DevTests'
